Update system_of_linear_equations.ipynb
Fixed an error in my equation (no full latex support) and fixed an error in the statement of the relation of singular vectors on the solution of Ax=b. It is only an approximation for the overdetermined system (not under-determined ones). For under-determined ones, all vectors corresponding to a singular value of zero are an exact solution.